+++ programs/drift/src/state/oracle_map.rs
@@ -88,7 +88,7 @@ impl<'a> OracleMap<'a> {
         oracle_id: &OracleIdentifier,
         last_oracle_price_twap: i64,
         max_confidence_interval_multiplier: u64,
-        slots_before_stale_for_amm_override: i64,
+        slots_before_stale_for_amm_override: i8,
     ) -> DriftResult<(&OraclePriceData, OracleValidity)> {
         if self.should_get_quote_asset_price_data(&oracle_id.0) {
             return Ok((&self.quote_asset_price_data, OracleValidity::Valid));
